계산시간 리턴되도록 수정

+++ src/main/java/com/green/secondproject/teacher/TeacherController.java
@@ -245,7 +245,7 @@ public class TeacherController {
             semester : 학기<br>
             midFinal : 중간(1), 기말(2)
             """)
-    public int calcRank(@RequestBody CalcClassRankParam p) {
+    public double calcRank(@RequestBody CalcClassRankParam p) {
         return service.calcRank(p);
     }
